requirements.txt
streamlit>=1.35.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.8
langchain-groq>=0.1.5
langgraph>=0.1.9
chromadb>=0.5.0
pypdf>=4.2.0
openai>=1.30.0
groq>=0.9.0
python-dotenv>=1.0.1
tiktoken>=0.7.0
pydantic>=2.7.0
